test: cover email scheduler
